Create README.md
Luxafar curates holidays that cater perfectly to this mind set, as they provide the freedom to explore and experience a destination on one’s own terms. From cooking with locals to exploring hidden trails, Luxafar offer experiences that allow travellers to feel a part of that place they visit rather than an outsider looking in.